+++ app/Http/Controllers/User/TicketController.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Http\Controllers\User;
+
+use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
+
+use App\Models\CustomUserTable;
+use App\Models\Ticket;
+use App\Models\Office;
+
+class TicketController extends Controller
+{
+    //
+}