+++ resources/views/user/ticket.blade.php
@@ -1,0 +1,18 @@
+<x-user-layout>
+  @slot('customCSS')
+
+  @endslot
+
+  @slot('modals')
+
+  @endslot
+
+  @slot('content')
+
+  @endslot
+
+  @slot('customJS')
+
+  <script src="{{ asset('js/ajax/user/ticket.ajax.js')}}" defer></script>
+  @endslot
+</x-user-layout>