+++ routes/user.php
@@ -11,7 +11,6 @@ Route::get('user/home', function () {
             // If the user is an admin, redirect to admin home page
             return redirect()->route('admin/home');
         } else {
-            // If the user is authenticated but not an admin, redirect to user home page
             return view('user/home');
         }
     } else {
@@ -19,3 +18,18 @@ Route::get('user/home', function () {
         return redirect()->route('login');
     }
 })->name('user/home');
+
+Route::get('user/ticket', function () {
+    if (Auth::check()) {
+        // Check if the user is logged in
+        if (Auth::user()->is_admin) {
+            // If the user is an admin, redirect to admin home page
+            return redirect()->route('admin/home');
+        } else {
+            return view('user/ticket');
+        }
+    } else {
+        // If the user is not authenticated, redirect to login page
+        return redirect()->route('login');
+    }
+})->name('user/ticket');